1.03.340.0 release, converted to a maven project, fixed issues
Fixed issues are: #26, #27, #28, #30, #32, #35, #36, #39

Note that #26, #27 required no changes, because Optional Structure
fields and Unions are purely serializer implementations and the standard
information model does not contain such types, therefore it is outside
the scope of the stack.

Note that most file locations changed in the process of converting to a
maven project (and some obsolete documentation that is no longer
required is removed), most of the relevant documentation is now merged
with the Readme.md to be more easily displayable in the github page.