Fixes for milestone version 1.03.341
Fixed GH issues
#18,#22,#45,#46,#47,#48,#49,#50,#52,#53,#60,#61,#65,#63,#66,#69.

Notes:
#22: See new classes in org.opcfoundation.ua.cert packages

#45: refactored class

#46, #48, refactored pom to have a jdk6 profile that is by default not
on (see #47)

#47: fixed via maven-javadoc-plugin:fix goal + manual edit for errors
that could not be fixed, this solution was chosen because it seems
profiles cannot be activated by toolchains, therefore an autoactivated
profile for jdk8 would activate when maven is run with 8 but the javadoc
would use 6 via toolchain and fail on unknown flag if the doclint:none
would be used. The quality of the autogenerated javadocs is what it is,
but at least now the project should be buildable using just jdk8.

As the expected build/development environment shall use 6 to compile, a
auto-jdk-8-activated profile would need to be disabled always. I would
prefer to have 0 'remember to activate' profile, but this is probably
the best workaround to allow easier contributions by people having only
jdk8 with 0 profiles activated and have just one profile 'jdk6' for the
"normal environment".

#49, 52: Changes applied to class files, codegenerator not affected yet

#50: Added support for StandardEngineeringUnits, see it's constants.
Note that the constant naming is a best effort of parsing the
description

#65: Structure interface now contains a .clone method to allow easier
cloning of compile-time-unknown Structure. Note that the
CloneNotSupportedException is not in the method signature as every
Structure is now expected to be cloneable (all existing structures have
been).

Also changed all structures to inherit from AbstractStructure that does
hide the CloneNotSupportedException override from Object signature to
allow existing Structures to be compliant with the method without
chainging their implementation to have the CloneNotSupportedException as
failing the .clone operation on a Structure is now an implementation
error.