#168, added Object.class mapping to use Variant encodings
Old BinaryDecoder.get had mappings for Object.class, but
EncoderUtils.put did not, so it was removed in the refactorings because
it did not make much sense given that it was only on the decoder side.

Putting it back and adding encoder side, as this might be useful if
Object.class is used as a field type for mapping BaseDataType. Using
Variant directly is an option, but that does not play well with
inheritance of BaseDataType-> other mapped UA types.