#136, created filters for the resource files, removed DecimalDataType id
To avoid fixing resource files manually in the future, there is now
ResourceFileFixerUtil.

DecimalDataType was removed from the Identifiers as it is not part of
the specification (the actual DecimalDataType "Structure" was removed
earlier as part of this issue). The filters include the actual 'Decimal'
type, because it should not be made to a Structure as it is mapped
directly to java BigDecimal and anyway uses custom encoding rules.