Fixed #121, Fixed #159, refactorings and changes related to SecurityMode
Created new class SecurityPolicyUri to hold constants needed for
SecurityPolicy and HttpsSecurityPolicy that are now enums.

Removed array constants from SecurityMode and moved them to
SecurityPolicy, avoids confusion what MessageSecurityModes the constants
had. SecurityModes can be now made with SecurityMode.combine that takes
a set of MessageSecurityModes and SecurityPolicies and creates suitable
combinations.

HTTPS now uses the scheme 'opc.https' as defined in the spec 1.04 part 6
section 7.4.1. Additionally UriUtil now uses port numbers defined in
7.6.

Removed logic in Server that would assume SecurityMode.NONE for HTTPS,
as Sign modes can be used to ensure application level authentication as
part of the CreateSession. Also added some methods that help to get this
information.

Removed HTTPS Cert client validation for the HttpsServer as per spec
1.04 Part 6 section 7.4.1 mentions that servers shall allow clients to
connect without providing a certificate during the negoation of the
HTTPS connection. NOTE! this means if the endpoint is created with
SecurityMode NONE, it will allow anyone to connect to the server (and if
used in conjunction Anonymous usertokenpolicy, to access it as well).
SecurityPolicies other than NONE should be used to prevent this.

#163 fix typo, AES128_SHA256_RSAOAEP should use aes128, not 256.

Changed: Defining nonce lengths directly in SecurityPolicy.